ai/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
pydantic
requests

paddlepaddle==2.6.2
paddleocr==2.7.3
numpy<2.0.0
opencv-python-headless
Pillow

ultralytics>=8.0.0
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.4.1+cpu
torchvision==0.19.1+cpu
torchaudio==2.4.1+cpu

groq
python-docx
pypdf